backend/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langchain
langchain-community
langchain-ollama
chromadb
pypdf